+++ build.sh
@@ -1,4 +1,8 @@
+podman run --rm \
+  -v "$(pwd)":/app \
+  -w /app \
+  maven:3.9-eclipse-temurin-8 mvn package -DskipTests
 
-mvn package -DskipTests || exit
+sudo chown -R "$(id -u):$(id -g)" target
 
-cp target/CommunicationMod.jar ~/Library/Application\ Support/Steam/steamapps/common/SlayTheSpire/SlayTheSpire.app/Contents/Resources/mods/
+cp target/HttpCommunicationMod.jar ~/Library/Application\ Support/Steam/steamapps/common/SlayTheSpire/SlayTheSpire.app/Contents/Resources/mods/