+++ src/main/java/httpcommunicationmod/OnStateChangeSubscriber.java
@@ -1,0 +1,7 @@
+package httpcommunicationmod;
+
+import basemod.interfaces.ISubscriber;
+
+public interface OnStateChangeSubscriber extends ISubscriber {
+    void receiveOnStateChange();
+}